+++ src/types/AuthToken.d.ts
@@ -1,0 +1,4 @@
+declare interface AuthToken {
+    user: string;
+    pass: string;
+}